feat: クリックで拡大表示可能な画像表示コンポーネントを追加
